requirements.txt
streamlit
langchain
python-dotenv
faiss-cpu
openai
tiktoken